rentIHA/requirements.txt
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
cssbeautifier==1.15.1
Django==5.0.4
django-environ==0.11.2
django-serverside-datatable==2.1.0
djangorestframework==3.15.1
EditorConfig==0.12.4
html-tag-names==0.1.2
html-void-elements==0.1.0
jsbeautifier==1.15.1
json5==0.9.25
pathspec==0.12.1
psycopg2==2.9.9
PyYAML==6.0.1
regex==2023.12.25
six==1.16.0
sqlparse==0.5.0
tqdm==4.66.4
tzdata==2024.1